+++ client/src/main.jsx
@@ -10,8 +10,7 @@ import PetsPage from './pages/PetsPage.jsx';
 import FurniturePage from './pages/FurniturePage.jsx';
 import SportsPage from './pages/SportsPage.jsx';
 import CartPage from './pages/CartPage.jsx';
-import ConfirmPage from './pages/ConfirmPage.jsx';
-import CheckoutForm from './pages/CheckoutForm.jsx';
+import CheckoutPage from './pages/CheckoutForm.jsx';
 
 
 const router = createBrowserRouter([
@@ -45,12 +44,8 @@ const router = createBrowserRouter([
         element: <CartPage />
       },
       {
-        path: '/confirm',
-        element: <ConfirmPage />
-      },
-      {
         path: '/checkout',
-        element: <CheckoutForm />
+        element: <CheckoutPage />
       }
     ]
   }
